refactor: split PersonaFields into per-field components, fix context menu layout
