+++ mods/persona/ui/src/index.css
@@ -740,10 +740,25 @@ body {
   }
 }
 
-/* Inline thumbnail for PersonaTab */
+/* Context menu persona tab header layout */
+.persona-tab-header {
+  display: flex;
+  gap: var(--hud-space-lg);
+  margin-bottom: var(--hud-space-xl);
+  align-items: flex-start;
+}
+
 .thumb-inline {
   width: 80px;
   aspect-ratio: 1 / 1;
-  float: left;
-  margin: 0 12px 12px 0;
+  flex-shrink: 0;
+  margin: 0;
+}
+
+.persona-tab-header-fields {
+  flex: 1;
+  min-width: 0;
+  display: flex;
+  flex-direction: column;
+  gap: var(--hud-space-xl);
 }
